requirements.txt
aws-cdk-lib==2.63.1
boto3==1.26.84
botocore==1.29.84
constructs>=10.0.0,<11.0.0